remove image_wip folder
This folder still contains all of the Chicago Face Database images
that haven't been cut down to the ones actually used in the pilot
task. They're not supposed to be redistributed, so they've been
removed from the task. Ideally this will free up a ton of
unnecessary bulk in the main task folder as well.