Update task readme
Updated the task readme file to include updated
details about the task design and remove notes
that are not publicly comprehensible. On the
to-do list is to figure out how to summarise
the task changes as it moves forward, but not
in such a way as to become cluttered/unhelpful.
TBD!